versionName "1.2.4" 尝试修复报错

+++ qiaopcplayer/src/main/cpp/MyAudio.cpp
@@ -336,48 +336,48 @@ int MyAudio::getCurrentSampleRateForOpensles(int sample_rate) {
     switch (sample_rate)
     {
         case 8000:
-            rate = SL_SAMPLINGRATE_8;
+            rate = (int)SL_SAMPLINGRATE_8;
             break;
         case 11025:
-            rate = SL_SAMPLINGRATE_11_025;
+            rate = (int)SL_SAMPLINGRATE_11_025;
             break;
         case 12000:
-            rate = SL_SAMPLINGRATE_12;
+            rate = (int)SL_SAMPLINGRATE_12;
             break;
         case 16000:
-            rate = SL_SAMPLINGRATE_16;
+            rate = (int)SL_SAMPLINGRATE_16;
             break;
         case 22050:
-            rate = SL_SAMPLINGRATE_22_05;
+            rate = (int)SL_SAMPLINGRATE_22_05;
             break;
         case 24000:
-            rate = SL_SAMPLINGRATE_24;
+            rate = (int)SL_SAMPLINGRATE_24;
             break;
         case 32000:
-            rate = SL_SAMPLINGRATE_32;
+            rate = (int)SL_SAMPLINGRATE_32;
             break;
         case 44100:
-            rate = SL_SAMPLINGRATE_44_1;
+            rate = (int)SL_SAMPLINGRATE_44_1;
             break;
         case 48000:
-            rate = SL_SAMPLINGRATE_48;
+            rate = (int)SL_SAMPLINGRATE_48;
             break;
         case 64000:
-            rate = SL_SAMPLINGRATE_64;
+            rate = (int)SL_SAMPLINGRATE_64;
             break;
         case 88200:
-            rate = SL_SAMPLINGRATE_88_2;
+            rate = (int)SL_SAMPLINGRATE_88_2;
             break;
         case 96000:
-            rate = SL_SAMPLINGRATE_96;
+            rate = (int)SL_SAMPLINGRATE_96;
             break;
         case 192000:
-            rate = SL_SAMPLINGRATE_192;
+            rate = (int)SL_SAMPLINGRATE_192;
             break;
         default:
-            rate =  SL_SAMPLINGRATE_44_1;
+            rate = (int)SL_SAMPLINGRATE_44_1;
     }
-    return (uint)rate;
+    return (int)rate;
 }
 
 void MyAudio::pause() {
